requirements.txt
flask
flask_sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
Flask-Migrate
